WIP: move tests back to /test and reference new /tmdDir
